% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_names.R
\name{fix_names}
\alias{fix_names}
\title{fix gene names to make them consistent between files}
\usage{
fix_names(data)
}
\arguments{
\item{data}{A express matrix with gene names as column names}
}
\value{
a data frame with gene names fixed
}
\description{
some genes have different names in different files of our demo data, make the names of these genes unique
}
\examples{
data<-fix_names(data)

}
