% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_integrate.R
\name{network_integrate}
\alias{network_integrate}
\title{Title}
\usage{
network_integrate(sc_knn, mc_knn, bulk_knn, mut_inf_mat1, mut_inf_mat2)
}
\arguments{
\item{mut_inf_mat2}{}
}
\description{
Title
}
